Move & clean Main views
