+++ ArchipelagoParser/Views/MainWindow.xaml
@@ -4,6 +4,8 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:NoNiDev.ArchipelagoParser.Views"
+        xmlns:rando="clr-namespace:NoNiDev.ArchipelagoParser.Views.RandoStat"
+        xmlns:sohar="clr-namespace:NoNiDev.ArchipelagoParser.Views.SOHAR"
         xmlns:uc="clr-namespace:NoNiDev.ArchipelagoParser.Views.CustomUserControl"
         mc:Ignorable="d"
         Title="Archipelago Parser" Height="720" Width="1280">
@@ -30,10 +32,10 @@
         <!-- Row 2-->
         <TabControl Grid.Row ="2" Margin="0,5,0,5">
             <TabItem Header="RandoStats">
-                <local:RandoStatsView />
+                <rando:RandoStatsView />
             </TabItem>
             <TabItem Header="SOHAR">
-                <local:SOHARView Loaded="SOHARView_Loaded"/>
+                <sohar:SOHARView Loaded="SOHARView_Loaded"/>
             </TabItem>
         </TabControl>
     </Grid>
